+++ parma_analytics/bl/register_measurement_values.py
@@ -1,5 +1,6 @@
 """This module contains the functions for registering measurement values."""
 import logging
+from dataclasses import dataclass
 from datetime import datetime
 from functools import partial
 from typing import Any
@@ -7,15 +8,14 @@ from typing import Any
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
 
-from parma_analytics.bl.generate_report import (
-    process_news_data,
-)
+from parma_analytics.bl.generate_report import GenerateNewsInput, generate_news
 from parma_analytics.db.prod.company_source_measurement_query import (
     create_company_measurement_query,
     get_by_company_and_measurement_ids_query,
 )
 from parma_analytics.db.prod.engine import get_engine, get_session
 from parma_analytics.db.prod.measurement_value_query import MeasurementValueCRUD
+from parma_analytics.db.prod.models.company_source_measurement import CompanyMeasurement
 from parma_analytics.db.prod.models.measurement_value_models import (
     MeasurementCommentValue,
     MeasurementDateValue,
@@ -27,10 +27,13 @@ from parma_analytics.db.prod.models.measurement_value_models import (
     MeasurementParagraphValue,
     MeasurementTextValue,
 )
+from parma_analytics.db.prod.models.news import News
 from parma_analytics.db.prod.reporting import get_users_subscribed_to_company
 from parma_analytics.reporting.gmail.email_service import EmailService
 from parma_analytics.reporting.news_comparison_engine import (
+    NewsComparisonEngineReturn,
     check_notification_rules,
+    create_news,
     get_source_module_id,
 )
 from parma_analytics.reporting.slack.send_slack_messages import SlackService
@@ -160,6 +163,56 @@ def handle_value(
         return measurement_id
     else:
         raise ValueError(f"Invalid measurement type: {measurement_type}")
+
+
+@dataclass
+class NewsInputData:
+    """News input data for process_news_data."""
+
+    company_id: int
+    source_measurement_id: int
+    data_source_id: int
+    company_measurement: CompanyMeasurement
+    value: Any
+    timestamp: datetime
+
+
+async def process_news_data(
+    comparison_engine_result: NewsComparisonEngineReturn, news_input_data: NewsInputData
+) -> None:
+    """Asynchoronously process the data and send notifications."""
+    # create news and send notifications, only if rules are satisfied
+    if comparison_engine_result.is_rules_satisfied:
+        try:
+            news_input = GenerateNewsInput(
+                company_id=news_input_data.company_id,
+                source_measurement_id=news_input_data.source_measurement_id,
+                company_measurement_id=news_input_data.company_measurement.company_measurement_id,
+                current_value=news_input_data.value,
+                trigger_change=comparison_engine_result.percentage_difference,
+                previous_value=comparison_engine_result.previous_value,
+                aggregation_method=comparison_engine_result.aggregation_method,
+            )
+            result = generate_news(news_input)
+        except SQLAlchemyError as e:
+            logger.error(f"A database error occurred while generating summary: {e}")
+        except Exception as e:
+            logger.error(f"An error occurred while generating news: {e}")
+
+        create_news(
+            News(
+                message=result["summary"],
+                company_id=news_input_data.company_id,
+                data_source_id=news_input_data.data_source_id,
+                trigger_factor=comparison_engine_result.percentage_difference,
+                title=result["title"],
+                timestamp=news_input_data.timestamp,
+                source_measurement_id=news_input_data.source_measurement_id,
+            )
+        )
+        send_notifications(
+            company_id=news_input_data.company_id, text=result["summary"]
+        )
 
 
 def send_notifications(company_id: int, text: str):
